+++ frontend/client/src/components/edit-Menu/EditMenuController.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import MenuDataGrid from '../components/edit-Menu/MenuDataGrid';
-import AddItemDialog from '../components/edit-Menu/AddItemDialog';
+import MenuDataGrid from './MenuDataGrid';
+import AddItemDialog from './AddItemDialog';
 import { Button } from '@material-ui/core';
 import DeleteIcon from '@material-ui/icons/Delete';
