+++ frontend/client/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App.js';
-import MenuTabController from './actions/MenuTabController';
-import StaffRecordsStage from './actions/StaffRecordsDataGridController';
+import MenuTabController from './components/edit-Menu/MenuTabController';
+import StaffRecordsStage from './components/staff-records/StaffRecordsDataGridController';
 
 ReactDOM.render(<App />, document.getElementById('root'));
